+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Change the EncodedChunk UUID generation (aka RaptorqHeader)
 - Change `raptorq` dependency from `1.6` to `2.0`
+- Change UDP sender to raise error if timeout`
 
 ## [0.6.1] - 2024-04-10